removing model operation from MatchController while moving it to Repository/MatchRepo.js
