+++ controllers/MatchControllers.js
@@ -1,47 +1,12 @@
-import sequelize from '../db.js'
+import {getMatchesInDb} from '../Repository/MatchRepo.js'
 
-const formatMatches = (results) => {
-
-    let obj = {
-        match_id: 0,
-        }
-  
-    const matches = []
-
-    const hash = results.reduce(
-        (p, c) => (p[c.matchId] ? p[c.matchId].push(c) : (p[c.matchId] = [c]), p),
-        {},
-    )
-
-    const newData = Object.keys(hash).map(k => ({matchId: k, pstats: hash[k]}))
-    
-    newData.map(details => {
-    obj = {
-        id: details.matchId,
-        created: details.pstats[0].createdAt,
-        map_name: details.pstats[0].map,
-        ct_score: details.pstats[0].ct_score,
-        t_score: details.pstats[0].t_score,
-        winner: details.pstats[0].m_winner,
-        players_details: details.pstats,
+export const getAllMatches = (req, res) => {
+  getMatchesInDb().then(matches => {
+    if (!matches) {
+      res.status(404).end()
     }
-    matches.push(obj)
-    })
-
-    return matches;
+    return res.send(matches)
+  })
 }
 
-export const getAllMatches = 
-    (req, res) => {
-        sequelize.query(
-          'SELECT * FROM matches INNER JOIN pstats ON matches.id = pstats.matchId',
-        )
-        .then(([results, metadata]) => {
-            return res.send(formatMatches(results))
-        })
-        .catch(err => {
-            res.status(500).end()
-        })
-      }
-
-export default getAllMatches;
+export default getAllMatches
